pyproject.toml
[project]
name = "cot-compression"
version = "0.1.0"
description = "财经领域长思维链压缩基线实现"
authors = [
    {name = "CoT-compression Team"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.36.0", 
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "requests>=2.28.0",
    "vllm>=0.2.0",
    "peft>=0.7.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.64.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=7.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]